fix(server): move enum key map handeling to utils and use it also in multiselect option
